pyproject.toml
[tool.poetry]
name = "nonebot-plugin-epicsteam"
version = "0.0.1"
description = ""
authors = ["fllesser <fllessive@gmail.com>"]
readme = "README.md"
packages = [{include = "nonebot_plugin_epicsteam"}]
repository = "https://github.com/fllesser"
keywords = ["nonebot", "nonebot2"]

[tool.poetry.dependencies]
python = "^3.10"

nonebot2 = "^2.4.0"
nonebot-adapter-onebot = "^2.4.6"
nonebot-plugin-alconna = "^0.54.2"
nonebot-plugin-localstore = "^0.7.3"
nonebot-plugin-apscheduler = "^0.5.0"

[tool.poetry.group.test.dependencies]
nb-cli = "^1.4.2"
nonebot2 = {version = "^2.4.0", extras = ["fastapi"]}

[tool.nonebot]
adapters = [{name="Onebot", module_name="nonebot.adapters.onebot.v11"}]
plugins = ["nonebot_plugin_epicsteam"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"